重构代码，并且将配置文件信息迁移到zk

+++ src/main/java/com/sohu/sns/monitor/bucket/TimeoutBucket.java
@@ -3,33 +3,34 @@ package com.sohu.sns.monitor.bucket;
 import com.google.common.base.Strings;
 
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 
 /**
- * 超时桶
+ * 超时信息统计桶
  * Created by Gary on 2015/10/19
  */
 public class TimeoutBucket {
     /**
      * 存放数据的alpha桶
      */
-    private static final ConcurrentHashMap<String, AtomicLong> bucketAlpha = new ConcurrentHashMap<String, AtomicLong>();
+    private static final ConcurrentHashMap<String, AtomicInteger> bucketAlpha = new ConcurrentHashMap<String, AtomicInteger>();
     /**
      * 存放数据的beta桶
      */
-    private static final ConcurrentHashMap<String,AtomicLong> bucketBeta = new ConcurrentHashMap<String, AtomicLong>();
+    private static final ConcurrentHashMap<String,AtomicInteger> bucketBeta = new ConcurrentHashMap<String, AtomicInteger>();
 
     /**
      * 正在工作中的桶
      */
-    private static ConcurrentHashMap<String,AtomicLong> bucket = bucketAlpha;
+    private static ConcurrentHashMap<String,AtomicInteger> bucket = bucketAlpha;
 
     /**
      * 切换桶
      * @return
      */
-    public static ConcurrentHashMap<String,AtomicLong> exchange() {
-        ConcurrentHashMap<String, AtomicLong> lastBucket = bucket;
+    public static ConcurrentHashMap<String,AtomicInteger> exchange() {
+        ConcurrentHashMap<String, AtomicInteger> lastBucket = bucket;
         if(bucket == bucketAlpha){
             bucket = bucketBeta;
         } else {
@@ -38,7 +39,7 @@ public class TimeoutBucket {
         return lastBucket;
     }
 
-    private static ConcurrentHashMap<String,AtomicLong> getBucket() {
+    private static ConcurrentHashMap<String,AtomicInteger> getBucket() {
         return bucket;
     }
 
@@ -50,8 +51,8 @@ public class TimeoutBucket {
         if (Strings.isNullOrEmpty(key)) {
             return;
         }
-        ConcurrentHashMap<String, AtomicLong> b = getBucket();
-        AtomicLong timeoutCount = b.get(key);
+        ConcurrentHashMap<String, AtomicInteger> b = getBucket();
+        AtomicInteger timeoutCount = b.get(key);
         if (null != timeoutCount) {
             timeoutCount.addAndGet(1);
         } else {
@@ -59,7 +60,7 @@ public class TimeoutBucket {
                 if(null != b.get(key)) {
                     insertData(key);
                 } else {
-                    AtomicLong countTemp = new AtomicLong(1);
+                    AtomicInteger countTemp = new AtomicInteger(1);
                     b.put(key, countTemp);
                 }
             }
